+++ src/app/authentication/signup-form/signup-form.component.ts
@@ -3,7 +3,7 @@ import { IResponse } from 'response-type';
 import { IUserForm, LoginResponse } from '../shared';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../environments/environment';
-import { UserService } from '@services/user.service';
+import { UserService, DataSyncGuard } from '@services/user.service';
 import { Router } from '@angular/router';
 import { GetNetworkError } from '@app/common';
 @Component({
@@ -26,9 +26,11 @@ export class SignupFormComponent implements OnInit {
     private http: HttpClient,
     private user: UserService,
     private router: Router,
+    private sync: DataSyncGuard
   ) { }
 
   ngOnInit() {
+    this.sync.UnSync();
   }
