+++ src/environments/environment.hmr.ts
@@ -1,5 +1,5 @@
 export const environment = {
-  production: false,
+  production: true,
   hmr: true,
   api: 'http://localhost:1337'
 };
